Made XCBConn poll method blocking
If <XCBConn as XConn>::poll_for_event() is non-blocking,
the main event loop starts to consume a lot of CPU time,
as it runs millions of polls every second.

To fix this, the method was temporarily made blocking.

Todo: determine whether this method should stay non-blocking,
and the event loop is modified, or if the method should
be changed to blocking.