overhauled platform-agnostic code
in order to use workspaces without having to pass around references,
each workspace now holds onto a handle to a RuntimeConfig - a
ManagerConfig. this allows us to activate and deactivate workspaces in a
WorkspaceMux without having to pass in references to a configuration.

the x11 dummy module was also deleted, as the dummy conn is no longer
relevant, since the call order was flipped (platform now calls manager
instead of manager calls config).